Removed unnecessary static_cast.

+++ src/mongo/db/pipeline/expression.h
@@ -955,7 +955,7 @@ public:
         : DateExpressionAcceptingTimeZone<ExpressionDayOfMonth>("$dayOfMonth", expCtx) {}
 
     Value evaluateDate(Date_t date, const TimeZone& timeZone) const final {
-        return Value(static_cast<int>(timeZone.dateParts(date).dayOfMonth));
+        return Value(timeZone.dateParts(date).dayOfMonth);
     }
 };
 
@@ -1112,7 +1112,7 @@ public:
         : DateExpressionAcceptingTimeZone<ExpressionHour>("$hour", expCtx) {}
 
     Value evaluateDate(Date_t date, const TimeZone& timeZone) const final {
-        return Value(static_cast<int>(timeZone.dateParts(date).hour));
+        return Value(timeZone.dateParts(date).hour);
     }
 };
 
@@ -1286,7 +1286,7 @@ public:
         : DateExpressionAcceptingTimeZone<ExpressionMillisecond>("$millisecond", expCtx) {}
 
     Value evaluateDate(Date_t date, const TimeZone& timeZone) const final {
-        return Value(static_cast<int>(timeZone.dateParts(date).millisecond));
+        return Value(timeZone.dateParts(date).millisecond);
     }
 };
 
@@ -1297,7 +1297,7 @@ public:
         : DateExpressionAcceptingTimeZone<ExpressionMinute>("$minute", expCtx) {}
 
     Value evaluateDate(Date_t date, const TimeZone& timeZone) const final {
-        return Value(static_cast<int>(timeZone.dateParts(date).minute));
+        return Value(timeZone.dateParts(date).minute);
     }
 };
 
@@ -1336,7 +1336,7 @@ public:
         : DateExpressionAcceptingTimeZone<ExpressionMonth>("$month", expCtx) {}
 
     Value evaluateDate(Date_t date, const TimeZone& timeZone) const final {
-        return Value(static_cast<int>(timeZone.dateParts(date).month));
+        return Value(timeZone.dateParts(date).month);
     }
 };
 
@@ -1472,7 +1472,7 @@ public:
         : DateExpressionAcceptingTimeZone<ExpressionSecond>("$second", expCtx) {}
 
     Value evaluateDate(Date_t date, const TimeZone& timeZone) const final {
-        return Value(static_cast<int>(timeZone.dateParts(date).second));
+        return Value(timeZone.dateParts(date).second);
     }
 };
 
@@ -1783,7 +1783,7 @@ public:
         : DateExpressionAcceptingTimeZone<ExpressionYear>("$year", expCtx) {}
 
     Value evaluateDate(Date_t date, const TimeZone& timeZone) const final {
-        return Value(static_cast<int>(timeZone.dateParts(date).year));
+        return Value(timeZone.dateParts(date).year);
     }
 };
